feat: 특정 유저가 올린 쇼츠 영상 리스트 받아오기 api 구현

+++ BackEnd/Chocolate/src/main/java/com/ssafy/chocolate/video/repository/ShortsJpaRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface ShortsJpaRepository extends JpaRepository<StreamShortClips, Integer> {
-    List<StreamShortClips> findByUserNo(int userNo);
+    List<StreamShortClips> findByUserNo(Long userNo);
     StreamShortClips findByStreamNo(Long streamNo);
     void deleteByStreamNo(Long streamNo);
 }
